Initializing colorblind agent
