+++ main.py
@@ -1,3 +1,4 @@
+from agents1.ColorblindAgent import ColorblindAgent
 from bw4t.BW4TWorld import BW4TWorld
 from bw4t.statistics import Statistics
 from agents1.BW4TBaselineAgent import BaseLineAgent
@@ -11,8 +12,8 @@ press the start button in god mode to start the session.
 
 if __name__ == "__main__":
     agents = [
-        {'name':'agent1', 'botclass':BaseLineAgent, 'settings':{'slowdown':10}},
-        {'name':'agent2', 'botclass':BaseLineAgent, 'settings':{}},
+        {'name':'agent1', 'botclass':ColorblindAgent, 'settings':{}},
+        # {'name':'agent2', 'botclass':BaseLineAgent, 'settings':{}},
         {'name':'human', 'botclass':Human, 'settings':{}}
         ]
